super commit (update + default)

+++ app/Http/Controllers/ProfileController.php
@@ -10,7 +10,8 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\View\View;
 use App\Models\Address;
-USE Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\DB;
+use App\Models\User;
 
 
 
@@ -96,23 +97,72 @@ class ProfileController extends Controller
         
         $address->save();
 
-        return Redirect::route('profile.edit')->with('status', 'profile-updated');
+        return Redirect::route('profile.edit')->with('status', 'profile-created');
 
     }
 
 
-    public function updateAddress(Request $request)
+    public function updateAddress($code , Request $request)
     {
+        $address = Address::where('code', '=', $code)->first();
 
+        $address->user_id = auth()->id();
+        $address->local_name = $request->local_name;
+        $address->street = $request->street;
+        $address->district = $request->district;
+        $address->state = $request->state;
+        $address->complement = $request->complement;
+        $address->zip_code = $request->zip_code;
+
+        $address->save();
+
+        return Redirect::route('profile.edit')->with('status', 'profile-updated');
+    }
+
+    public function showUpdateAddress(Request $request)
+    {
+        $code = $request->code;
+        $user_id = Auth::user()->id;
+        $address = Address::where('user_id',$user_id)->get();
+        $address_code = Address::where('code', $code)->first();
+
+        if($user_id && $address && $address_code){
+            return view('profile.address-form-update' , [
+                'code' => $code,
+                'address' => $address_code,
+            ]);
+
+        }
+
+        
     }
 
     public function destroyAddress(Request $request)
     {
 
-        DB::table('addresses')->where('code','=',$request->code)->delete();
-        
+        /* 
+            Função de deletar endereço com mais uma camada de segurança. 
+            Essa função busca o id do usuário autenticado no site e compara se esse id é daquele endereço mesmo, 
+            logo depois deleta, caso não mosta um erro.
 
-        return redirect()->route('profile.edit');
+        */
+
+
+        $user_id = Auth::user()->id;
+        $address_user_id = Address::where('user_id',$user_id)->first();
+        
+        if($address_user_id){
+
+            DB::table('addresses')->where('code','=',$request->code)->delete();
+
+            return redirect()->route('profile.edit');
+
+        }else{
+
+            echo 'Desculpe, ocorreu um erro.';
+        }
+        
+        
     }
 
 
@@ -121,6 +171,31 @@ class ProfileController extends Controller
     {
         return view('profile.address-form');
     }
+
+
+    public function updateDefaultAddress($code)
+    {
+
+        $user_id = Auth::user()->id;
+
+        
+        $default = Address::where('code',$code)->first();
+
+        
+        
+        if(Auth::user() && $default->is_default == false){
+
+            Address::where('user_id',$user_id)->first()->where('is_default',true)->update(['is_default' => false]);
+
+
+            Address::where('code',$code)->first()->update(['is_default' => true]);
+
+        }
+        
+        return redirect()->back()->with('success', 'Default address updated successfully.');
+    }
+
+    
 
 
 }
